% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_helpers.R
\name{add_normalized_assays}
\alias{add_normalized_assays}
\title{Add normalized assays}
\usage{
add_normalized_assays(dds, which = c("CPM", "TPM", "FPKM"))
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{which}{Subset of \code{c("CPM","TPM","FPKM")} to add.}
}
\value{
The \code{DESeqDataSet} with the requested assays attached.
}
\description{
CPM is always computed; TPM/FPKM only when \code{\link[=has_feature_length]{has_feature_length()}} is TRUE
(otherwise skipped with a warning).
}
