% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsdashboard-package.R
\docType{package}
\name{ddsdashboard-package}
\alias{ddsdashboard}
\alias{ddsdashboard-package}
\title{ddsdashboard: Explore and Process RNA-seq DESeqDataSet Objects}
\description{
A Shiny dashboard for exploring, manipulating, and visualizing bulk RNA-seq results stored in DESeqDataSet (and convertible SingleCellExperiment) objects. Aimed at users with little to no bioinformatics background: quality control, metadata editing, filtering, normalization, dimensionality reduction, DESeq2 differential expression, heatmaps, and export. Also exposes reusable helper functions (normalization, feature lookup) for other projects.
}
\author{
\strong{Maintainer}: Siwat Chad Ruangroengkulrith \email{chorchatty@gmail.com}

Authors:
\itemize{
  \item Siwat Chad Ruangroengkulrith \email{chorchatty@gmail.com}
}

}
\keyword{internal}
