% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_helpers.R
\name{de_contrast_levels}
\alias{de_contrast_levels}
\title{Levels of a design variable (for contrast test/control pickers)}
\usage{
de_contrast_levels(dds, var)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{var}{A colData column name.}
}
\value{
Character vector of levels (factor order preserved; else sorted uniques).
}
\description{
Levels of a design variable (for contrast test/control pickers)
}
