% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_helpers.R
\name{default_sizefactor_config}
\alias{default_sizefactor_config}
\title{Default size-factor normalization config}
\usage{
default_sizefactor_config()
}
\value{
\code{list(control, custom_ids, type, provenance)}.
}
\description{
Default size-factor normalization config
}
