% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_helpers.R
\name{estimate_size_factors_endogenous}
\alias{estimate_size_factors_endogenous}
\title{Estimate size factors on endogenous genes}
\usage{
estimate_size_factors_endogenous(dds)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}
}
\value{
The \code{DESeqDataSet} with \code{sizeFactors} set (endogenous \code{controlGenes}).
}
\description{
Estimate size factors on endogenous genes
}
