% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_helpers.R
\name{import_gtf}
\alias{import_gtf}
\title{Import a GTF/GFF file}
\usage{
import_gtf(path)
}
\arguments{
\item{path}{Path to the annotation file.}
}
\value{
A \code{GRanges}.
}
\description{
Thin wrapper over \code{rtracklayer::import()} (accepts \code{.gtf} / \code{.gtf.gz}).
}
