% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_helpers.R
\name{palette_qualitative_names}
\alias{palette_qualitative_names}
\title{Names of the built-in qualitative palettes (back-compat shim)}
\usage{
palette_qualitative_names()
}
\value{
Character vector of qualitative palette names.
}
\description{
Names of the built-in qualitative palettes (back-compat shim)
}
