% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_helpers.R
\name{remove_coldata_column}
\alias{remove_coldata_column}
\title{Remove a colData column (design columns are protected)}
\usage{
remove_coldata_column(dds, name)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{name}{Column to remove.}
}
\value{
The updated \code{DESeqDataSet}.
}
\description{
Remove a colData column (design columns are protected)
}
