% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_helpers.R
\name{rename_samples}
\alias{rename_samples}
\title{Rename samples (colnames), enforcing uniqueness}
\usage{
rename_samples(dds, old, new)
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{old}{Existing sample name(s).}

\item{new}{Replacement name(s), same length as \code{old}.}
}
\value{
The updated \code{DESeqDataSet}.
}
\description{
Rename samples (colnames), enforcing uniqueness
}
