% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the dds dashboard}
\usage{
run_app(...)
}
\arguments{
\item{...}{Passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\value{
A Shiny app object (invisibly when run interactively).
}
\description{
Launch the dds dashboard
}
