% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_helpers.R
\name{set_feature_class}
\alias{set_feature_class}
\title{Set the feature_class of selected features (bulk)}
\usage{
set_feature_class(dds, ids, class = c("endogenous", "spike_in", "exogenous"))
}
\arguments{
\item{dds}{A \code{DESeqDataSet}.}

\item{ids}{Feature ids (rownames) to tag.}

\item{class}{One of \code{endogenous} / \code{spike_in} / \code{exogenous}.}
}
\value{
The updated \code{DESeqDataSet} (with \code{feature_class} ensured).
}
\description{
Set the feature_class of selected features (bulk)
}
