% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_dds}
\alias{state_dds}
\title{Reactive read of the current working \code{dds}}
\usage{
state_dds(state)
}
\arguments{
\item{state}{An app-state object from \code{\link[=new_app_state]{new_app_state()}}.}
}
\value{
A reactive expression yielding the working \code{dds} (or \code{NULL}).
}
\description{
Reactive read of the current working \code{dds}
}
