% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_reset}
\alias{state_reset}
\title{Restore the originally loaded object}
\usage{
state_reset(state)
}
\arguments{
\item{state}{App-state object.}
}
\value{
\code{state}, invisibly.
}
\description{
Restore the originally loaded object
}
