% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_undo}
\alias{state_undo}
\title{Undo the most recent edit}
\usage{
state_undo(state)
}
\arguments{
\item{state}{App-state object.}
}
\value{
\code{state}, invisibly.
}
\description{
Undo the most recent edit
}
