% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_helpers.R
\name{palette_qualitative_names}
\alias{palette_qualitative_names}
\title{Names of the built-in qualitative palettes}
\usage{
palette_qualitative_names()
}
\value{
Character vector of palette names accepted by \code{\link[=palette_qualitative]{palette_qualitative()}}
and \code{\link[=palette_discrete]{palette_discrete()}}.
}
\description{
Names of the built-in qualitative palettes
}
